Get Latest Exchange Rates For Devises

Returns a data.frame with the latest exchange rate(s) for the given currencies. Default will convert to USD ($).

Usage

get_changes(from = NULL, to = "USD", .verbose = T, ...)

Arguments

from,

default = NULL A character string representing the base currency (e.g., "USD"). The user must provide from values with one of the following method :

passing a paired character list, e.g., c("EUR" = "USD", "RON" = "EUR") ;

or passing one or several from values :

associated with a single to value, e.g., from = c('EUR', "RON"), to = "USD"

or associated with a list of from values of the same length, e.g., from = c('EUR', "RON"), to = c("USD", "EUR") .

to,

default = "USD" A character string representing the target currencies, e.g., c("EUR", "USD").

.verbose

logical If TRUE, messages are displayed, e.g., when invalid symbols are detected. Default is FALSE.

...

Other parameters passed to get_a_change - internal mecanism for get_changes_historic

Value

The returned dataframe contains daily exchange rates, and the following columns:

currency

character - The base currency.

symbol

character - The Yahoo Finance symbol (e.g., "EURUSD=X").

exchangename

character - The exchange place name, i.e. 'CCY' for currencies.

intrumenttype

character - The type of financial instrument, supposed to be 'CURRENCY'.

firsttradedate

POSIXct - The oldest date available on the Yahoo Finance API for an exchange rates historic.

gmtoffset

numeric - The difference with GMT time (seconds).

timezone

character - The market's timezone.

exchangetimezonename

character - The name of the market's timezone, e.g., 'Europe/London'.

regularmarketprice

numeric - The latest market price.

regularmarketdayhigh

numeric - The market highest price of this day.

regularmarketdaylow

numeric - The market lowest price of this day.

fiftytwoweekhigh

numeric - The highest price in the last 52 weeks.

fiftytwoweeklow

numeric - The lowest price in the last 52 weeks.

previousclose

numeric - The last closing price.

regular_timezone

character - The regular market's timezone.

from

character - The currency converted into another, e.g., if the from value is 1$ ('USD'), you want to receive a certain amount of the other currency to reach 1$.

to

character - The currency that you want to convert into : all the numeric values (not integer) of the row are expressed with this currency.

timestamp

POSIXct - The corresponding date (YYYY-MM-DD).

Details

Return a data.frame with additionnal attributes:

n.currencies

integer - Number of unique currencies in the data.frame.

currencies

character - A vector of currency symbols in the data.frame (e.g., "BTC", "ETH", "USD").

exchange

character - A vector of exchange pairs (e.g., "BTC => USD"). If not applicable, NULL (no entry).

date

Date - The date when the data was retrieved, set using Sys.Date().

crypto

logical - TRUE if cryptocurrencies are present, otherwise FALSE.

References

Source : https://query2.finance.yahoo.com/v8/finance/chart/

See also

Examples

# Latest exchange rates from € and ¥ to $ (default convert to 'USD")
df <- get_changes(from = c("EUR", "JPY"))
str(df)
#> 'data.frame':	2 obs. of  18 variables:
#>  $ currency            : chr  "USD" "USD"
#>  $ symbol              : chr  "EURUSD=X" "JPYUSD=X"
#>  $ exchangename        : chr  "CCY" "CCY"
#>  $ instrumenttype      : chr  "CURRENCY" "CURRENCY"
#>  $ firsttradedate      : POSIXct, format: "2003-12-01" "1996-10-30"
#>  $ gmtoffset           : int  3600 3600
#>  $ timezone            : chr  "BST" "BST"
#>  $ exchangetimezonename: chr  "Europe/London" "Europe/London"
#>  $ regularmarketprice  : num  1.0963 0.0068
#>  $ regularmarketdayhigh: num  1.111 0.0069
#>  $ regularmarketdaylow : num  1.0929 0.0068
#>  $ fiftytwoweekhigh    : num  1.1212 0.0072
#>  $ fiftytwoweeklow     : num  1.0184 0.0062
#>  $ previousclose       : num  1.1052 0.0069
#>  $ regular_timezone    : chr  "BST" "BST"
#>  $ from                : chr  "EUR" "JPY"
#>  $ to                  : chr  "USD" "USD"
#>  $ timestamp           : POSIXct, format: "2025-04-04 21:29:36" "2025-04-04 20:59:41"
#>  - attr(*, "n.currencies")= int 3
#>  - attr(*, "currencies")= chr [1:3] "EUR" "JPY" "USD"
#>  - attr(*, "exchange")= chr [1:2] "EUR => USD" "JPY => USD"
#>  - attr(*, "date")= Date[1:1], format: "2025-04-06"
#>  - attr(*, "crypto")= logi FALSE
# User could pass a named list of character
df2 <- get_changes(from = c("EUR" = "RON", "USD" = "EUR"))
# Or pass paired values as 2 list (equivalent to hereabove line) :
same_as_df2 <- get_changes(from = c("EUR", "USD"), to =c("RON" , "EUR"))