Get Latest Cryptocurrencies Values (USD) and Market Data

Return a data.frame of latest financial data for 100 cryptocurrencies, e.g., actual values of each cryptocurrency in USD.

Usage

get_crypto(keep = NULL, .verbose = T)

Arguments

keep

A character vector of symbols to filter the results. If NULL (default), no filtering is applied, and all available indices are returned.

.verbose

logical, default = TRUE. If TRUE, send messages to the console if necessary.

Value

A data frame with 100 observations and 13 variables:

symbol

character - Cryptocurrency ticker symbol.

name

character - Name of the cryptocurrency, along with its quoted 'real-world' currency.

price

numeric - Current price of the cryptocurrency in USD ($).

change

numeric - Absolute price change in USD since the last closing of the exchange place.

change_percent

numeric - Percentage price change since the last closing of the exchange place (%).

market_cap

numeric - Total market capitalization of the cryptocurrency.

volume

numeric - 24-hour trading volume.

volume_in_currency_24hr

numeric - 24-hour trading volume in the associated 'real-world' currency (most of the time redundant with the volume column).

total_volume_all_currencies_24hr

numeric - 24-hour total trading volume across all currency pairs.

circulating_supply

numeric - Total circulating supply of the cryptocurrency.

X52_wk_change_percent

numeric - Percentage change in price over the last 52 weeks (%).

from

character, the currency converted into another, e.g., if the from value is 1$ ('USD'), you want to receive a certain amount of the other currency to reach 1$.

to

character, the currency that you want to convert into : all the numeric values (not integer) in this line of the data.frame are expressed with this currency.

Details

Return a data.frame with additionnal attributes:

n.currencies

integer - Number of unique currencies in the data.frame.

currencies

character - A vector of currency symbols in the data.frame (e.g., "BTC", "ETH", "USD").

exchange

character - A vector of exchange pairs (e.g., "BTC => USD"). If not applicable, NULL (no entry).

date

Date - The date when the data was retrieved, set using Sys.Date().

crypto

logical - TRUE if cryptocurrencies are present, otherwise FALSE.

References

Source : Yahoo Finance 'crypto' page, https://finance.yahoo.com/markets/crypto/all/

Examples

krypto <- get_crypto()
head(krypto)
#>     symbol            name        price      change change_percent  market_cap
#> 1  BTC-USD     Bitcoin USD 85260.980000 2926.550000           3.55 1.69200e+12
#> 2  ETH-USD    Ethereum USD  1916.200000   92.450000           5.07 2.31208e+11
#> 3 USDT-USD Tether USDt USD     0.999879    0.000048           0.00 1.43976e+11
#> 4  XRP-USD         XRP USD     2.147200    0.063700           3.06 1.24980e+11
#> 5  BNB-USD         BNB USD   611.830000    7.880000           1.31 8.71680e+10
#> 6  SOL-USD      Solana USD   126.940000    1.750000           1.40 6.50570e+10
#>       volume volume_in_currency_24hr total_volume_all_currencies_24hr
#> 1 2.8137e+10              2.8137e+10                       2.8137e+10
#> 2 1.5162e+10              1.5162e+10                       1.5162e+10
#> 3 5.8265e+10              5.8265e+10                       5.8265e+10
#> 4 4.0610e+09              4.0610e+09                       4.0610e+09
#> 5 2.0860e+09              2.0860e+09                       2.0860e+09
#> 6 3.0510e+09              3.0510e+09                       3.0510e+09
#>   circulating_supply X52_wk_change_percent from  to
#> 1        1.98450e+07                 26.13  BTC USD
#> 2        1.20660e+08                -44.36  ETH USD
#> 3        1.43993e+11                 -0.05 USDT USD
#> 4        5.82060e+10                256.96  XRP USD
#> 5        1.42472e+08                  9.88  BNB USD
#> 6        5.12505e+08                -31.17  SOL USD