Find Similar Financial Symbols

Given symbol(s), retrieve identical symbols (according to Yahoo Finance) and a score of similarity.

Usage

get_similar(symbols = "SAAB-B.ST", .verbose = F, ...)

Arguments

symbols

character A character string representing a financial symbol to search.

.verbose

logical If TRUE, messages are displayed when invalid symbols are detected. Default is TRUE.

...

Other symbols (char. or list of char.)

Value

A data.frame with the symbols associated with those provided by the user and similarity scores, according to Yahoo Finance.

from

character - Financial symbol provided by the user.

symbol

character - Symbol associated with the other 'from' symbol.

score

numeric - Similarity score, according to Yahoo Finance.

References

Source : https://query2.finance.yahoo.com/v6/finance/quote/validate

Examples

get_similar(symbols =   "AAPL,GOOGL")
#>          from symbol    score
#> AAPL.1   AAPL   AMZN 0.294944
#> AAPL.2   AAPL   TSLA 0.278101
#> AAPL.3   AAPL   GOOG 0.261226
#> AAPL.4   AAPL   META 0.250701
#> AAPL.5   AAPL   MSFT 0.210578
#> GOOGL.1 GOOGL   BABA 0.121617
#> GOOGL.2 GOOGL   NFLX 0.120005
#> GOOGL.3 GOOGL   PYPL 0.118386
#> GOOGL.4 GOOGL      V 0.118228
#> GOOGL.5 GOOGL   META 0.117542
get_similar(symbols =   c("AAPL", "GOOGL"))
#>          from symbol    score
#> AAPL.1   AAPL   AMZN 0.294944
#> AAPL.2   AAPL   TSLA 0.278101
#> AAPL.3   AAPL   GOOG 0.261226
#> AAPL.4   AAPL   META 0.250701
#> AAPL.5   AAPL   MSFT 0.210578
#> GOOGL.1 GOOGL   BABA 0.121617
#> GOOGL.2 GOOGL   NFLX 0.120005
#> GOOGL.3 GOOGL   PYPL 0.118386
#> GOOGL.4 GOOGL      V 0.118228
#> GOOGL.5 GOOGL   META 0.117542