Retrieve latest financial data of symbols

Calls an internal API to fetch flat latest (non-historical) metadata for one or more asset symbols. Returns a data.frame with one row per valid financial symbol, with information such as latest prices and trading context.

Usage

get_values(symbols = c("AAPL", "GOOGL"), .verbose = F)

Arguments

symbols

Character vector. One or more asset symbols (e.g., c("AAPL", "GOOGL")).

.verbose

Logical. If TRUE, displays verbose output during the fetching process. Default is FALSE.

Value

A data.frame with 1 row per symbol provided and 21 columns, including:

currency

character - Trading currency of the asset (e.g., "USD").

symbol

character - Symbol of the asset, e.g., ticker symbol.

exchangename

character - Abbreviated exchange place or financial instrument name.

fullexchangename

character - Full name of the exchange place or financial instrument.

instrumenttype

character - Type of financial instrument (e.g., "EQUITY").

firsttradedate

POSIXct - Datetime when the asset was first traded.

regularmarkettime

POSIXct - Timestamp of the latest regular market quote.

hasprepostmarketdata

logical - Indicates if pre/post-market data is available.

gmtoffset

integer - Offset from GMT in seconds.

timezone

character - Abbreviated timezone name of the market.

exchangetimezonename

character - Timezone name of the exchange location.

regularmarketprice

numeric - Latest regular market trading price.

fiftytwoweekhigh

numeric - highest price over a 52-weeks period.

fiftytwoweeklow

numeric - lowest price over a 52-weeks period.

regularmarketdayhigh

numeric - Highest price during the current market day.

regularmarketdaylow

numeric - Lowest price during the current market day.

regularmarketvolume

integer - Volume traded during the current market day.

shortname

character - Shortened or common name of the asset.

chartpreviousclose

numeric - Closing price shown in charts.

previousclose

numeric - Previous official market close price.

Details

Return a data.frame with additionnal attributes:

n.currencies

integer - Number of unique currencies in the data.frame.

currencies

character - A vector of currency symbols in the data.frame (e.g., "BTC", "ETH", "USD").

exchange

character - A vector of exchange pairs (e.g., "BTC => USD"). If not applicable, NULL (no entry).

date

Date - The date when the data was retrieved, set using Sys.Date().

crypto

logical - TRUE if cryptocurrencies are present, otherwise FALSE.

References

Source : https://query2.finance.yahoo.com/v8/finance/chart/

Examples

get_values(c("AAPL", "GOOGL","SAAB-B.ST"))
#>   currency    symbol exchangename fullexchangename instrumenttype
#> 1      USD      AAPL          NMS         NasdaqGS         EQUITY
#> 2      USD     GOOGL          NMS         NasdaqGS         EQUITY
#> 3      SEK SAAB-B.ST          STO        Stockholm         EQUITY
#>        firsttradedate   regularmarkettime hasprepostmarketdata gmtoffset
#> 1 1980-12-12 14:30:00 2025-04-04 20:00:02                 TRUE    -14400
#> 2 2004-08-19 13:30:00 2025-04-04 20:00:03                 TRUE    -14400
#> 3 2000-01-03 08:00:00 2025-04-04 15:29:55                FALSE      7200
#>   timezone exchangetimezonename regularmarketprice fiftytwoweekhigh
#> 1      EDT     America/New_York             188.38           260.10
#> 2      EDT     America/New_York             145.60           207.05
#> 3     CEST     Europe/Stockholm             379.40           420.00
#>   fiftytwoweeklow regularmarketdayhigh regularmarketdaylow regularmarketvolume
#> 1          164.08               199.88             187.345           125910913
#> 2          145.38               151.05             145.380            62259539
#> 3          203.70               420.00             365.000             6985601
#>        shortname chartpreviousclose previousclose
#> 1     Apple Inc.             203.19        203.19
#> 2  Alphabet Inc.             150.72        150.72
#> 3 SAAB AB ser. B             411.65        411.65