Search Financial Asset From a Keyword (Single Text)

Searches for financial data based on a free-text texts and retrieve asset symbol(s), name(s), exchange, and asset type.

Usage

search_asset_quick(
  texts = "Dow Jones",
  .verbose = TRUE,
  region = NULL,
  lang = "en"
)

Arguments

texts

character string representing the search texts. This can be a company name, index, or financial term. Default is "Dow Jones".

.verbose

logical If TRUE, print additional details about the search process. Default is TRUE.

region

A character string specifying the region for the search (e.g., "US", "EU"). Default is NULL, meaning no region filter is applied.

lang

A character string specifying the language of the data returned. Default is "en" for English.

Value

A data.frame with assets symbols and names.

symbol: The financial symbol (e.g., stock ticker or index).

name: The full name of the financial entity (e.g., 'Dow Jones Industrial Average').

exch: The exchange on which the symbol is listed (e.g., 'DJI', 'CBT').

type: The type of the financial instrument (e.g., I for Index, F for Futures, E for Exchange Traded Fund).

exchdisp: The exchange name displayed (e.g., 'Dow Jones').

typedisp: A long name for the type of financial instrument (e.g., 'Index', 'Futures' or 'ETF' for 'Exchange Traded Fund').

Examples

if (FALSE) { # \dontrun{
# Example of searching for financial data related to "Dow Jones"
results <- search_asset_quick(texts = "Dow Jones")
} # }