Get Latest Financial Information and Ticker Symbol From Company Name(s)

Get latest overall insights and ticker symbols associated with free texts, such as companies names. Several ticker symbols associated with various exchanges places could be returned, with , companies names, last price on the marketplace, sector/category (if available), etc.

Usage

search_from_text(texts, .verbose = F, exchange = NULL, quotetype = NULL)

Arguments

texts

character A character string or a list of character representing the text(s) to search for, e.g., company names.

.verbose

logical, default = FALSE If TRUE, messages are displayed, e.g., when invalid symbols are detected.

exchange

(optionnal) A character string representing the exchange place(s) to consider - exact match, e.g., 'STO' (Stockholm stock exchange). Default keep all results.

quotetype

(optionnal) A character string representing the type of entry to search for, e.g., a stock market ('ETF', 'EQUITY' or 'MUTUALFUND') or a major indice ('INDEX').

Value

A data frame with the following columns:

symbol

numeric - The ticker symbol associated with an entry.

exchange

character - .

shortname

character - .

quotetype

numeric - .

index

character - .

score

numeric - .

typedisp

character - .

longname

character - .

exchdisp

character - .

isyahoofinance

character - .

searched

character - .

Details

Return a data.frame with additionnal attributes:

n.currencies

integer - Number of unique currencies in the data.frame.

currencies

character - A vector of currency symbols in the data.frame (e.g., "BTC", "ETH", "USD").

exchange

character - A vector of exchange pairs (e.g., "BTC => USD"). If not applicable, NULL (no entry).

date

Date - The date when the data was retrieved, set using Sys.Date().

crypto

logical - TRUE if cryptocurrencies are present, otherwise FALSE.

References

Source : https://query2.finance.yahoo.com/v1/

Examples

oil <- search_from_text(texts = c("TOTAL", "SHELL", "BP"),  quotetype = "EQUITY")
#Get data on marketplace(s)
swedish <- search_from_text(c("SAAB", "VOLVO"),  exchange = c("STO", "PAR"))