Get Latest Financial Information and Ticker Symbol From Company Name(s)

Get latest overall insights and ticker symbols associated with free texts, such as companies texts. Several ticker symbols associated with various exchanges places could be returned, with , companies texts, last price on the marketplace, sector/category (if available), etc.

Usage

search_summary(texts, exchange = NULL, sector = NULL)

Arguments

texts

A character string representing the company name to search for.

exchange

(optionnal) A character string representing the exchange place(s) to consider (exact match). Default keep all the exchange places.

sector

(optionnal) A character string representing the sector(s) to consider (exact match). Default keep all results.

Value

A data frame with columns:

symbol

character - The ticker symbol associated with an asset, e.g., "VTI", "^DWCPF".

name

character - Name of the asset, e.g., ETF or index name.

last_price

numeric - The last price of the asset.

sector

character - The sector or industry category if available, e.g., 'Industrials', 'Consumer Cyclical'.

type

character - The type of asset (certainly "stocks").

exchange

character - The stock exchange place for this asset, e.g., 'PAR' is Paris (FR) exchange place.

searched

character - The text searched on Yahoo, e.g., "Dow Jones".

Details

Return a data.frame with additionnal attributes:

n.currencies

integer - Number of unique currencies in the data.frame.

currencies

character - A vector of currency symbols in the data.frame (e.g., "BTC", "ETH", "USD").

exchange

character - A vector of exchange pairs (e.g., "BTC => USD"). If not applicable, NULL (no entry).

date

Date - The date when the data was retrieved, set using Sys.Date().

crypto

logical - TRUE if cryptocurrencies are present, otherwise FALSE.

References

Source : search on https://finance.yahoo.com/lookup/

See also

Examples

oil <- search_summary(texts = c("Dow Jones", "Euronext"))
#Get data on marketplace(s)
cars <- search_summary(texts = c("RENAULT", "VOLVO"),  exchange = c("STO", "PAR"))