Retrieve latest financial data

Get latest insights, given financial symbols. Data such as latest prices and trading context are returned.

Usage

get_values(symbols = c("AAPL", "GOOGL"), .verbose = F)

Arguments

symbols

Character vector. One or more asset symbols (e.g., c("AAPL", "GOOGL")).

.verbose

Logical. If TRUE, displays verbose output during the fetching process. Default is FALSE.

Value

A data.frame with 1 row per valid financial symbol and 21 columns, including:

currency

character - Trading currency of the asset (e.g., "USD").

symbol

character - Symbol of the asset, e.g., ticker symbol.

exchangename

character - Abbreviated exchange place or financial instrument name.

fullexchangename

character - Full name of the exchange place or financial instrument.

instrumenttype

character - Type of financial instrument (e.g., "EQUITY").

firsttradedate

POSIXct - Datetime when the asset was first traded.

regularmarkettime

POSIXct - Timestamp of the latest regular market quote.

hasprepostmarketdata

logical - Indicates if pre/post-market data is available.

gmtoffset

integer - Offset from GMT in seconds.

timezone

character - Abbreviated timezone name of the market.

exchangetimezonename

character - Timezone name of the exchange location.

regularmarketprice

numeric - Latest regular market trading price.

fiftytwoweekhigh

numeric - Highest price over a 52-weeks period.

fiftytwoweeklow

numeric - Lowest price over a 52-weeks period.

regularmarketdayhigh

numeric - Highest price during the current market day.

regularmarketdaylow

numeric - Lowest price during the current market day.

regularmarketvolume

integer - Volume traded during the current market day.

shortname

character - Shortened or common name of the asset.

chartpreviousclose

numeric - Closing price shown in charts.

previousclose

numeric - Previous official market close price.

Details

Return a data.frame with additional attributes:

n.currencies

integer - Number of unique currencies in the data.frame.

currencies

character - A vector of currency symbols in the data.frame (e.g., "BTC", "ETH", "USD").

exchange

character - A vector of exchange pairs (e.g., "BTC => USD"). If not applicable, NULL (no entry).

date.fetch

Date - The date when the data was retrieved, set using Sys.Date().

crypto

logical - TRUE if cryptocurrencies are present, otherwise FALSE.

date.begin

POSIXct - The oldest observation in the dataset.

date.end

POSIXct - The most recent observation in the dataset.

date.dif

numeric - The number of seconds between date.begin and date.end, equivalent of difftime value.

See vignette("Functions_summary", package = "financr")

References

Source : https://query2.finance.yahoo.com/v8/finance/chart/

Examples

get_values(c("AAPL", "GOOGL","SAAB-B.ST"))
#>   currency    symbol exchangename fullexchangename instrumenttype
#> 1      USD      AAPL          NMS         NasdaqGS         EQUITY
#> 2      USD     GOOGL          NMS         NasdaqGS         EQUITY
#> 3      SEK SAAB-B.ST          STO        Stockholm         EQUITY
#>        firsttradedate   regularmarkettime hasprepostmarketdata gmtoffset
#> 1 1980-12-12 14:30:00 2025-10-02 20:00:01                 TRUE    -14400
#> 2 2004-08-19 13:30:00 2025-10-02 20:00:01                 TRUE    -14400
#> 3 2000-01-03 08:00:00 2025-10-03 13:08:37                FALSE      7200
#>   timezone exchangetimezonename regularmarketprice fiftytwoweekhigh
#> 1      EDT     America/New_York             257.13            260.1
#> 2      EDT     America/New_York             245.69            256.0
#> 3     CEST     Europe/Stockholm             573.70            587.5
#>   fiftytwoweeklow regularmarketdayhigh regularmarketdaylow regularmarketvolume
#> 1          169.21               258.18             254.150            42459369
#> 2          140.53               246.81             242.305            25343292
#> 3          203.70               575.40             564.000              907487
#>        shortname chartpreviousclose previousclose
#> 1     Apple Inc.             255.45        255.45
#> 2  Alphabet Inc.             244.90        244.90
#> 3 SAAB AB ser. B             566.40        566.40