Get Latest Cryptocurrencies Values (USD) and Market Data

Return a data.frame of latest financial data for 100 cryptocurrencies, e.g., actual values of each cryptocurrency in USD.

Usage

last_crypto(keep = NULL, .verbose = T)

Arguments

keep

character - Character vector of symbols to filter the results (perl expression, ignoring case). If NULL (default), no filtering is applied, and all available indices are returned.

.verbose

logical, default = TRUE - If TRUE, send messages to the console if necessary.

Value

A data frame with 100 observations and 13 variables:

symbol

character - Cryptocurrency ticker symbol.

name

character - Name of the cryptocurrency, along with its quoted 'real-world' currency.

price

numeric - Current price of the cryptocurrency in USD ($).

change

numeric - Absolute price change in USD since the last closing of the exchange place.

change_percent

numeric - Percentage price change since the last closing of the exchange place (%).

market_cap

numeric - Total market capitalization of the cryptocurrency.

volume

numeric - 24-hour trading volume.

volume_in_currency_24hr

numeric - 24-hour trading volume in the associated 'real-world' currency (most of the time redundant with the volume column).

total_volume_all_currencies_24hr

numeric - 24-hour total trading volume across all currency pairs.

circulating_supply

numeric - Total circulating supply of the cryptocurrency.

x52_wk_change_percent

numeric - Percentage change in price over the last 52 weeks (%).

from

character - The currency converted into another, e.g., if the from value is 1$ ('USD'), you want to receive a certain amount of the other currency to reach 1$.

to

character - The currency that you want to convert into : all the currency-related numeric values in this line of the data.frame are expressed with this currency.

Details

Return a data.frame with additional attributes:

n.currencies

integer - Number of unique currencies in the data.frame.

currencies

character - A vector of currency symbols in the data.frame (e.g., "BTC", "ETH", "USD").

exchange

character - A vector of exchange pairs (e.g., "BTC => USD"). If not applicable, NULL (no entry).

date.fetch

Date - The date when the data was retrieved, set using Sys.Date().

crypto

logical - TRUE if cryptocurrencies are present, otherwise FALSE.

date.begin

POSIXct - The oldest observation in the dataset.

date.end

POSIXct - The most recent observation in the dataset.

date.dif

numeric - The number of seconds between date.begin and date.end, equivalent of difftime value.

See vignette("Functions_summary", package = "financr")

References

Source : Yahoo Finance 'crypto' page, https://finance.yahoo.com/markets/crypto/all/

See also

Examples

krypto <- last_crypto()
head(krypto)
#>     symbol            name       price    change change_percent  market_cap
#> 1  BTC-USD     Bitcoin USD 120368.0900 1433.8300           1.21 2.39900e+12
#> 2  ETH-USD    Ethereum USD   4479.3800   89.2800           2.03 5.40675e+11
#> 3  XRP-USD         XRP USD      3.0392    0.0580           1.95 1.81964e+11
#> 4 USDT-USD Tether USDt USD      1.0007    0.0002           0.02 1.76360e+11
#> 5  BNB-USD         BNB USD   1109.2200   54.6700           5.18 1.54387e+11
#> 6  SOL-USD      Solana USD    230.7600    4.4700           1.98 1.25819e+11
#>        volume volume_in_currency_24hr total_volume_all_currencies_24hr
#> 1 7.39990e+10             7.39990e+10                      7.39990e+10
#> 2 4.78530e+10             4.78530e+10                      4.78530e+10
#> 3 7.53700e+09             7.53700e+09                      7.53700e+09
#> 4 1.67864e+11             1.67864e+11                      1.67864e+11
#> 5 4.54300e+09             4.54300e+09                      4.54300e+09
#> 6 8.99100e+09             8.99100e+09                      8.99100e+09
#>   circulating_supply x52_wk_change_percent x52_wk_range1 x52_wk_range2 from  to
#> 1        1.99280e+07                 94.44      58895.21     124457.12  BTC USD
#> 2        1.20703e+08                 85.85       1386.80       4953.73  ETH USD
#> 3        5.98720e+10                468.80          0.49          3.65  XRP USD
#> 4        1.76241e+11                  0.06          1.00          1.00 USDT USD
#> 5        1.39185e+08                 96.38        509.84       1112.70  BNB USD
#> 6        5.45238e+08                 64.04         96.59        294.33  SOL USD