Get Latest Cryptocurrencies Values (USD) and Market Data

Return a data.frame of latest financial data for 100 cryptocurrencies, e.g., actual values of each cryptocurrency in USD.

Usage

last_crypto(keep = NULL, .verbose = T)

Arguments

keep

character - Character vector of symbols to filter the results (perl expression, ignoring case). If NULL (default), no filtering is applied, and all available indices are returned.

.verbose

logical, default = TRUE - If TRUE, send messages to the console if necessary.

Value

A data frame with 100 observations and 13 variables:

symbol

character - Cryptocurrency ticker symbol.

name

character - Name of the cryptocurrency, along with its quoted 'real-world' currency.

price

numeric - Current price of the cryptocurrency in USD ($).

change

numeric - Absolute price change in USD since the last closing of the exchange place.

change_percent

numeric - Percentage price change since the last closing of the exchange place (%).

market_cap

numeric - Total market capitalization of the cryptocurrency.

volume

numeric - 24-hour trading volume.

volume_in_currency_24hr

numeric - 24-hour trading volume in the associated 'real-world' currency (most of the time redundant with the volume column).

total_volume_all_currencies_24hr

numeric - 24-hour total trading volume across all currency pairs.

circulating_supply

numeric - Total circulating supply of the cryptocurrency.

x52_wk_change_percent

numeric - Percentage change in price over the last 52 weeks (%).

from

character - The currency converted into another, e.g., if the from value is 1$ ('USD'), you want to receive a certain amount of the other currency to reach 1$.

to

character - The currency that you want to convert into : all the currency-related numeric values in this line of the data.frame are expressed with this currency.

Details

Return a data.frame with additional attributes:

n.currencies

integer - Number of unique currencies in the data.frame.

currencies

character - A vector of currency symbols in the data.frame (e.g., "BTC", "ETH", "USD").

exchange

character - A vector of exchange pairs (e.g., "BTC => USD"). If not applicable, NULL (no entry).

date.fetch

Date - The date when the data was retrieved, set using Sys.Date().

crypto

logical - TRUE if cryptocurrencies are present, otherwise FALSE.

date.begin

POSIXct - The oldest observation in the dataset.

date.end

POSIXct - The most recent observation in the dataset.

date.dif

numeric - The number of seconds between date.begin and date.end, equivalent of difftime value.

See vignette("Functions_summary", package = "financr")

References

Source : Yahoo Finance 'crypto' page, https://finance.yahoo.com/markets/crypto/all/

See also

Examples

krypto <- last_crypto()
head(krypto)
#>     symbol            name        price      change change_percent  market_cap
#> 1  BTC-USD     Bitcoin USD 76851.450000 -1.6256e+03          -2.07 1.52500e+12
#> 2  ETH-USD    Ethereum USD  1467.250000 -8.7710e+01          -5.64 1.77062e+11
#> 3 USDT-USD Tether USDt USD     0.998946 -4.0500e-04          -0.04 1.44040e+11
#> 4  XRP-USD         XRP USD     1.818300 -8.6200e-02          -4.53 1.05964e+11
#> 5  BNB-USD         BNB USD   552.280000 -2.5200e+00          -0.45 7.86840e+10
#> 6 USDC-USD    USD Coin USD     1.000000  1.0000e-04           0.01 6.01580e+10
#>       volume volume_in_currency_24hr total_volume_all_currencies_24hr
#> 1 4.8928e+10              4.8928e+10                       4.8928e+10
#> 2 2.0323e+10              2.0323e+10                       2.0323e+10
#> 3 8.1360e+10              8.1360e+10                       8.1360e+10
#> 4 6.4600e+09              6.4600e+09                       6.4600e+09
#> 5 1.6680e+09              1.6680e+09                       1.6680e+09
#> 6 1.5087e+10              1.5087e+10                       1.5087e+10
#>   circulating_supply x52_wk_change_percent from  to
#> 1        1.98490e+07                 14.60  BTC USD
#> 2        1.20676e+08                -55.63  ETH USD
#> 3        1.44192e+11                  0.00 USDT USD
#> 4        5.82750e+10                209.17  XRP USD
#> 5        1.42471e+08                 -4.54  BNB USD
#> 6        6.01560e+10                  0.02 USDC USD