Search Financial Asset From Keywords (Multiple Texts)

Get ticker symbols associated with free texts, such as companies names. Several ticker symbols associated with various exchanges places could be returned, with companies names, sector/category (if available), etc.

Usage

search_assets(
  texts,
  exchange = NULL,
  type = NULL,
  add_latest_values = F,
  .verbose = F
)

Arguments

texts

character A character string or a list of character representing the text(s) to search for, e.g., company names.

exchange

character (optionnal) - A character string representing the exchange place(s) to consider - exact match, e.g., 'STO' (Stockholm stock exchange). Default keep all results.

type

(optionnal) A character string representing a type of asset to search for ('quotetype' column value is used to filter out results). Case insensitive.

Type	Description

ETF	Exchange-Traded Fund - A basket of stocks or bonds traded like a stock.

EQUITY	Equity - A stock representing ownership in a company.

MUTUALFUND	Mutual Fund - A pooled investment fund.

INDEX	Index - A benchmark representing a group of stocks (e.g., S&P 500).

FUTURE	Future - A contract to buy/sell an asset at a future date and price.

add_latest_values

logical, default = FALSE - If TRUE, search for the values of the symbols and add columns from get_values()

.verbose

logical, default = FALSE If TRUE, messages are displayed, e.g., when invalid symbols are detected.

Value

A data frame with the following columns:

symbol

character - The ticker symbol associated with an asset, e.g., "VTI", "^DWCPF".

shortname

character - A short name for the asset, e.g., ETF or index name.

longname

character - The full name of the asset, such as the full ETF or index name - sometimes not returned by Yahoo.

exchange

character - The exchange where the asset is listed, e.g., "PCX", "DJI", "NGM".

exchdisp

character - The full name of the exchange place where the asset is traded, e.g., "NYSEArca", "Dow Jones".

quoteType

character - The type of asset, e.g., "FUTURES", "INDEX".

typeDisp

character - The type of asset, formatted for display, e.g., "Futures", "Index").

score

numeric - A numerical score assigned by Yahoo in order to indicate the relevance of the matched result, i.e. similarity with the text.

isYahooFinance

logical - Indicates whether the symbol is recognized by Yahoo Finance - should always be TRUE in this context.

searched

character - The text searched on the Yahoo API, e.g., "Dow Jones".

Details

Return a data.frame with additionnal attributes:

n.currencies

integer - Number of unique currencies in the data.frame.

currencies

character - A vector of currency symbols in the data.frame (e.g., "BTC", "ETH", "USD").

exchange

character - A vector of exchange pairs (e.g., "BTC => USD"). If not applicable, NULL (no entry).

date.fetch

Date - The date when the data was retrieved, set using Sys.Date().

crypto

logical - TRUE if cryptocurrencies are present, otherwise FALSE.

date.begin

POSIXct - The oldest obs. in the dataset.

date.end

POSIXct - The most recent date in the dataset.

date.dif

numeric & difftime - The number of seconds between date.begin and date.end.

References

Source : https://query2.finance.yahoo.com/v1/

See also

Examples

indices <- search_assets(texts = c("Dow jones", "euronext"), type = "index" )

swed <- search_assets(c("VOLVO car", "RENAULT"),  exchange = c("STO", "PAR"))
head(swed)
#>        symbol           shortname             longname exchange  exchdisp
#> 1 VOLCAR-B.ST Volvo Car AB ser. B Volvo Car AB (publ.)      STO Stockholm
#> 9      RNO.PA             RENAULT           Renault SA      PAR     Paris
#>   quotetype typedisp score isyahoofinance  searched
#> 1    EQUITY   Equity 20004           TRUE volvo car
#> 9    EQUITY   Equity 20006           TRUE   renault

swed_last_values <- search_assets(c("VOLVO car", "SAAB")
,  exchange = "STO", add_latest_values = TRUE )

str(swed_last_values)
#> 'data.frame':	1 obs. of  28 variables:
#>  $ symbol              : chr "VOLCAR-B.ST"
#>  $ shortname           : chr "Volvo Car AB ser. B"
#>  $ longname            : chr "Volvo Car AB (publ.)"
#>  $ exchange            : chr "STO"
#>  $ exchdisp            : chr "Stockholm"
#>  $ quotetype           : chr "EQUITY"
#>  $ typedisp            : chr "Equity"
#>  $ score               : num 20004
#>  $ isyahoofinance      : logi TRUE
#>  $ searched            : chr "volvo car"
#>  $ currency            : chr "SEK"
#>  $ exchangename        : chr "STO"
#>  $ fullexchangename    : chr "Stockholm"
#>  $ instrumenttype      : chr "EQUITY"
#>  $ firsttradedate      : POSIXct, format: "2021-10-29 07:00:00"
#>  $ regularmarkettime   : POSIXct, format: "2025-04-04 15:29:38"
#>  $ hasprepostmarketdata: logi FALSE
#>  $ gmtoffset           : int 7200
#>  $ timezone            : chr "CEST"
#>  $ exchangetimezonename: chr "Europe/Stockholm"
#>  $ regularmarketprice  : num 17.7
#>  $ fiftytwoweekhigh    : num 44.8
#>  $ fiftytwoweeklow     : num 17.2
#>  $ regularmarketdayhigh: num 18.4
#>  $ regularmarketdaylow : num 17.2
#>  $ regularmarketvolume : int 10297687
#>  $ chartpreviousclose  : num 17.8
#>  $ previousclose       : num 17.8
#>  - attr(*, "n.currencies")= int 1
#>  - attr(*, "currencies")= chr "SEK"
#>  - attr(*, "date.fetch")= Date[1:1], format: "2025-04-06"
#>  - attr(*, "crypto")= logi FALSE
#>  - attr(*, "date.begin")= POSIXct[1:1], format: "2025-04-04 15:29:38"
#>  - attr(*, "date.end")= POSIXct[1:1], format: "2025-04-04 15:29:38"
#>  - attr(*, "date.dif")= 'difftime' int 0
#>   ..- attr(*, "units")= chr "secs"